Switch over to RepeatableArgumentMatcher to again allow getters
