--- src/main/java/org/openrewrite/java/migrate/apache/commons/lang/MethodInvocationMatcher.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.java.migrate.apache.commons.lang;
-
-import org.openrewrite.java.template.Matcher;
-import org.openrewrite.java.tree.Expression;
-import org.openrewrite.java.tree.J;
-
-class MethodInvocationMatcher implements Matcher<Expression> {
-    @Override
-    public boolean matches(Expression expr) {
-        return expr instanceof J.MethodInvocation;
-    }
-}
